Use src folder and air config file
