+++ .air.toml
@@ -1,0 +1,6 @@
+root = "./src/"
+tmp_dir = "tmp"
+
+[build]
+cmd = "go build -o ./src/tmp/bootstrap.exe src/KanjiServer.go"
+bin = "./src/tmp/bootstrap.exe"
