--- KanjiDBLib/JLPT Kanji/N3
@@ -1,370 +1,0 @@
-与
-両
-乗
-予
-争
-互
-亡
-交
-他
-付
-件
-任
-伝
-似
-位
-余
-例
-供
-便
-係
-信
-倒
-候
-値
-偉
-側
-偶
-備
-働
-優
-光
-全
-共
-具
-内
-冷
-処
-列
-初
-判
-利
-到
-制
-刻
-割
-加
-助
-努
-労
-務
-勝
-勤
-化
-単
-危
-原
-参
-反
-収
-取
-受
-可
-号
-合
-向
-君
-否
-吸
-吹
-告
-呼
-命
-和
-商
-喜
-回
-因
-困
-園
-在
-報
-増
-声
-変
-夢
-太
-夫
-失
-好
-妻
-娘
-婚
-婦
-存
-宅
-守
-完
-官
-定
-実
-客
-害
-容
-宿
-寄
-富
-寒
-寝
-察
-対
-局
-居
-差
-市
-師
-席
-常
-平
-幸
-幾
-座
-庭
-式
-引
-当
-形
-役
-彼
-徒
-得
-御
-必
-忘
-忙
-念
-怒
-怖
-性
-恐
-恥
-息
-悲
-情
-想
-愛
-感
-慣
-成
-戦
-戻
-所
-才
-打
-払
-投
-折
-抜
-抱
-押
-招
-指
-捕
-掛
-探
-支
-放
-政
-敗
-散
-数
-断
-易
-昔
-昨
-晩
-景
-晴
-暗
-暮
-曲
-更
-最
-望
-期
-未
-末
-束
-杯
-果
-格
-構
-様
-権
-横
-機
-欠
-次
-欲
-歯
-歳
-残
-段
-殺
-民
-求
-決
-治
-法
-泳
-洗
-活
-流
-浮
-消
-深
-済
-渡
-港
-満
-演
-点
-然
-煙
-熱
-犯
-状
-猫
-王
-現
-球
-産
-由
-申
-留
-番
-疑
-疲
-痛
-登
-皆
-盗
-直
-相
-眠
-石
-破
-確
-示
-礼
-祖
-神
-福
-科
-程
-種
-積
-突
-窓
-笑
-等
-箱
-米
-精
-約
-組
-経
-給
-絵
-絶
-続
-緒
-罪
-置
-美
-老
-耳
-職
-育
-背
-能
-腹
-舞
-船
-良
-若
-苦
-草
-落
-葉
-薬
-術
-表
-要
-規
-覚
-観
-解
-記
-訪
-許
-認
-誤
-説
-調
-談
-論
-識
-警
-議
-負
-財
-貧
-責
-費
-資
-賛
-越
-路
-身
-辞
-込
-迎
-返
-迷
-追
-退
-逃
-途
-速
-連
-進
-遅
-遊
-過
-達
-違
-遠
-適
-選
-部
-都
-配
-酒
-閉
-関
-阪
-降
-限
-除
-険
-陽
-際
-雑
-難
-雪
-静
-非
-面
-靴
-頂
-頭
-頼
-顔
-願
-類
-飛
-首
-馬
-髪
-鳴
