+++ data/JLPT Kanji/N4
@@ -1,0 +1,167 @@
+不
+世
+主
+事
+京
+仕
+代
+以
+会
+住
+体
+作
+使
+借
+元
+兄
+公
+写
+冬
+切
+別
+力
+勉
+動
+医
+去
+口
+古
+台
+同
+味
+品
+員
+問
+図
+地
+堂
+場
+売
+夏
+夕
+多
+夜
+妹
+姉
+始
+字
+安
+室
+家
+少
+屋
+工
+帰
+広
+店
+度
+建
+弟
+強
+待
+心
+思
+急
+悪
+意
+手
+持
+教
+文
+料
+新
+方
+旅
+族
+早
+明
+映
+春
+昼
+曜
+有
+服
+朝
+業
+楽
+歌
+止
+正
+歩
+死
+注
+洋
+海
+漢
+牛
+物
+特
+犬
+理
+用
+田
+町
+画
+界
+病
+発
+的
+目
+真
+着
+知
+研
+社
+私
+秋
+究
+空
+立
+答
+紙
+終
+習
+考
+者
+肉
+自
+色
+花
+英
+茶
+親
+言
+計
+試
+買
+貸
+質
+赤
+走
+起
+足
+転
+近
+送
+通
+週
+運
+道
+重
+野
+銀
+開
+院
+集
+青
+音
+題
+風
+飯
+飲
+館
+駅
+験
+魚
+鳥
+黒
