+++ data/JLPT Kanji/N5
@@ -1,0 +1,80 @@
+一
+七
+万
+三
+上
+下
+中
+九
+二
+五
+人
+今
+休
+何
+先
+入
+八
+六
+円
+出
+分
+前
+北
+十
+千
+午
+半
+南
+友
+右
+名
+四
+国
+土
+外
+大
+天
+女
+子
+学
+小
+山
+川
+左
+年
+後
+日
+時
+書
+月
+木
+本
+来
+東
+校
+母
+毎
+気
+水
+火
+父
+生
+男
+白
+百
+聞
+行
+西
+見
+話
+語
+読
+車
+金
+長
+間
+雨
+電
+食
+高
